fixed commit() to filter out inserted nodes that have later been deleted added 'newEntity' and 'updateEntity' commit names switched to module syntax added tests for changes.mjs
