+++ package.json
@@ -4,7 +4,7 @@
   "description": "REST api for curriculum data, uses the GraphQL server in https://github.com/slonl/curriculum-graphql-server. It is written in javascript, using nodejs, with the express framework.",
   "main": "src/api-server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "tap test/*.mjs",
     "start": "node src/api-server.js"
   },
   "repository": {
@@ -27,5 +27,8 @@
     "request-promise-native": "^1.0.9",
     "uuid": "^9.0.0",
     "watch": "^1.0.2"
+  },
+  "devDependencies": {
+    "tap": "^19.0.2"
   }
 }
