+++ www/templates/scripts.html
@@ -23,7 +23,7 @@
 <script src="/assets/js/api.js"></script>
 <script src="https://unpkg.com/@muze-nl/jsontag@0.9.2/dist/browser.js"></script>
 <script src="/assets/js/local-api.js"></script>
-<script src="/assets/js/changes.js"></script>
+<script src="/assets/js/changes.js" type="module"></script>
 <script src="/assets/js/spreadsheet.js"></script>
 <script src="/assets/js/sloDocument.js"></script>
 <script src="/assets/js/transformers.js"></script>
